[testsuite] Split Objective-C new syntax test
This splits the second longest test into separate test cases. Similar to
what we did for the Objective-C data formatters in r357786.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@357824 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 5fb8e37c609350d9c1b4c41ae6cc08a174488eaa)
